requirements.txt
Cython==0.29.21
easydict==1.9
lmdb==1.0.0
notebook==6.1.5
numba==0.56.4
numpy==1.19.2
opencv-python==4.7.0.72
pandas==1.1.5
Pillow==8.0.1
pycparser==2.20
Pygments==2.7.3
pyparsing==2.4.7
msgpack==1.0.4
pytz==2020.4
PyWavelets==1.3.0
pyzmq==20.0.0
scikit-image==0.18.0
scikit-learn==1.0.2
scipy==1.5.4
seaborn==0.11.2
Send2Trash==1.5.0
six==1.15.0
sympy==1.10.1
torch==1.7.1
torchvision==0.8.2
tornado==6.1
tqdm==4.54.1
